+++ guillotina/db/strategies/simple.py
@@ -16,7 +16,8 @@ class SimpleStrategy(BaseStrategy):
 
     async def tpc_begin(self):
         await self.retrieve_tid()
-        await self._storage.start_transaction(self._transaction)
+        if self._transaction._db_txn is None:
+            await self._storage.start_transaction(self._transaction)
 
     async def tpc_vote(self):
         if not self.writable_transaction: